requirements.txt
streamlit
openai
pydantic
python-dotenv

